toggle inventory on 'E' press

+++ src/gfx/renderer.c
@@ -7,41 +7,45 @@ void renderer_init(struct Renderer *self) {
     self->current_shader = SHADER_NONE;
 
     self->shaders[SHADER_BASIC_TEXTURE] = shader_create(
-        "res/shaders/basic_texture.vs", "res/shaders/basic_texture.fs",
-        2, (struct VertexAttr[]) {
-            { .index = 0, .name = "position" },
-            { .index = 1, .name = "uv" }
-        });
+            "res/shaders/basic_texture.vs", "res/shaders/basic_texture.fs",
+            2, (struct VertexAttr[]) {
+                    {.index = 0, .name = "position"},
+                    {.index = 1, .name = "uv"}
+            }
+    );
 
     self->shaders[SHADER_CHUNK] = shader_create(
-        "res/shaders/chunk.vs", "res/shaders/chunk.fs",
-        3, (struct VertexAttr[]) {
-            { .index = 0, .name = "position" },
-            { .index = 1, .name = "uv" },
-            { .index = 2, .name = "color" }
-        });
+            "res/shaders/chunk.vs", "res/shaders/chunk.fs",
+            3, (struct VertexAttr[]) {
+                    {.index = 0, .name = "position"},
+                    {.index = 1, .name = "uv"},
+                    {.index = 2, .name = "color"}
+            }
+    );
 
     self->shaders[SHADER_SKY] = shader_create(
-        "res/shaders/sky.vs", "res/shaders/sky.fs",
-        2, (struct VertexAttr[]) {
-            { .index = 0, .name = "position" },
-            { .index = 1, .name = "uv" }
-        });
+            "res/shaders/sky.vs", "res/shaders/sky.fs",
+            2, (struct VertexAttr[]) {
+                    {.index = 0, .name = "position"},
+                    {.index = 1, .name = "uv"}
+            }
+    );
 
     self->shaders[SHADER_BASIC_COLOR] = shader_create(
-        "res/shaders/basic_color.vs", "res/shaders/basic_color.fs",
-        1, (struct VertexAttr[]) {
-            { .index = 0, .name = "position" }
-        });
+            "res/shaders/basic_color.vs", "res/shaders/basic_color.fs",
+            1, (struct VertexAttr[]) {
+                    {.index = 0, .name = "position"}
+            }
+    );
 
 
     self->block_atlas = blockatlas_create(
-        "res/images/blocks.png",
-        (ivec2s) {{ 16, 16 }}
+            "res/images/blocks.png",
+            (ivec2s) {{16, 16}}
     );
 
     self->textures[TEXTURE_CROSSHAIR] = texture_create_from_path("res/images/crosshair.png");
-    self->textures[TEXTURE_CLOUDS] = texture_create_from_path("res/images/clouds.png"); 
+    self->textures[TEXTURE_CLOUDS] = texture_create_from_path("res/images/clouds.png");
     self->textures[TEXTURE_STAR] = texture_create_from_path("res/images/star.png");
     self->textures[TEXTURE_SUN] = texture_create_from_path("res/images/sun.png");
     self->textures[TEXTURE_MOON] = texture_create_from_path("res/images/moon.png");
@@ -53,8 +57,10 @@ void renderer_init(struct Renderer *self) {
 
     perspective_camera_init(&self->perspective_camera, radians(75.0f));
     ortho_camera_init(
-        &self->ortho_camera, GLMS_VEC2_ZERO,
-        (vec2s) {{ state.window->size.x, state.window->size.y }});
+            &self->ortho_camera,
+            GLMS_VEC2_ZERO,
+            (vec2s) {{(f32) state.window->size.x, (f32) state.window->size.y}}
+    );
 }
 
 void renderer_destroy(struct Renderer *self) {
@@ -76,8 +82,10 @@ void renderer_prepare(struct Renderer *self, enum RenderPass pass) {
     switch (pass) {
         case PASS_2D:
             ortho_camera_init(
-                &self->ortho_camera, GLMS_VEC2_ZERO,
-                (vec2s) {{ state.window->size.x, state.window->size.y }});
+                    &self->ortho_camera,
+                    GLMS_VEC2_ZERO,
+                    (vec2s) {{(f32) state.window->size.x, (f32) state.window->size.y}}
+            );
             glClear(GL_DEPTH_BUFFER_BIT);
             glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
             glDisable(GL_DEPTH_TEST);
@@ -87,8 +95,11 @@ void renderer_prepare(struct Renderer *self, enum RenderPass pass) {
             break;
         case PASS_3D:
             glClearColor(
-                self->clear_color.x, self->clear_color.y,
-                self->clear_color.z, self->clear_color.w);
+                    self->clear_color.x,
+                    self->clear_color.y,
+                    self->clear_color.z,
+                    self->clear_color.w
+            );
             glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
             glPolygonMode(GL_FRONT_AND_BACK, self->flags.wireframe ? GL_LINE : GL_FILL);
             glEnable(GL_DEPTH_TEST);
@@ -141,22 +152,25 @@ void renderer_use_shader(struct Renderer *self, enum ShaderType shader) {
 }
 
 void renderer_quad_color(
-    struct Renderer *self, vec2s size,
-    vec4s color, mat4s model) {
+        struct Renderer *self,
+        vec2s size,
+        vec4s color,
+        mat4s model
+) {
     renderer_use_shader(self, SHADER_BASIC_COLOR);
     renderer_set_view_proj(self);
     shader_uniform_mat4(self->shader, "m", model);
     shader_uniform_vec4(self->shader, "color", color);
 
     vbo_buffer(self->vbo, (f32[]) {
-        0, 0, 0,
-        0, size.y, 0,
-        size.x, size.y, 0,
-        size.x, 0, 0,
+            0, 0, 0,
+            0, size.y, 0,
+            size.x, size.y, 0,
+            size.x, 0, 0,
     }, 0, (4 * 3) * sizeof(f32));
 
     vbo_buffer(self->ibo, (u32[]) {
-        3, 0, 1, 3, 1, 2
+            3, 0, 1, 3, 1, 2
     }, 0, 6 * sizeof(u32));
 
     vao_attr(self->vao, self->vbo, 0, 3, GL_FLOAT, 0, 0);
@@ -167,10 +181,14 @@ void renderer_quad_color(
 }
 
 void renderer_quad_texture(
-    struct Renderer *self, struct Texture texture,
-    vec2s size, vec4s color,
-    vec2s uv_min, vec2s uv_max,
-    mat4s model) {
+        struct Renderer *self,
+        struct Texture texture,
+        vec2s size,
+        vec4s color,
+        vec2s uv_min,
+        vec2s uv_max,
+        mat4s model
+) {
     renderer_use_shader(self, SHADER_BASIC_TEXTURE);
     renderer_set_view_proj(self);
     shader_uniform_mat4(self->shader, "m", model);
@@ -178,19 +196,19 @@ void renderer_quad_texture(
     shader_uniform_vec4(self->shader, "color", color);
 
     vbo_buffer(self->vbo, (f32[]) {
-        0, 0, 0,
-        0, size.y, 0,
-        size.x, size.y, 0,
-        size.x, 0, 0,
+            0, 0, 0,
+            0, size.y, 0,
+            size.x, size.y, 0,
+            size.x, 0, 0,
 
-        uv_min.x, uv_min.y,
-        uv_min.x, uv_max.y,
-        uv_max.x, uv_max.y,
-        uv_max.x, uv_min.y
+            uv_min.x, uv_min.y,
+            uv_min.x, uv_max.y,
+            uv_max.x, uv_max.y,
+            uv_max.x, uv_min.y
     }, 0, ((4 * 3) + (4 * 2)) * sizeof(f32));
 
     vbo_buffer(self->ibo, (u32[]) {
-        3, 0, 1, 3, 1, 2
+            3, 0, 1, 3, 1, 2
     }, 0, 6 * sizeof(u32));
 
     vao_attr(self->vao, self->vbo, 0, 3, GL_FLOAT, 0, 0);
@@ -202,33 +220,37 @@ void renderer_quad_texture(
 }
 
 void renderer_aabb(
-    struct Renderer *self, AABB aabb, vec4s color,
-    mat4s model, enum FillMode fill_mode) {
+        struct Renderer *self,
+        AABB aabb,
+        vec4s color,
+        mat4s model,
+        enum FillMode fill_mode
+) {
     renderer_use_shader(self, SHADER_BASIC_COLOR);
     renderer_set_view_proj(self);
     shader_uniform_mat4(self->shader, "m", model);
     shader_uniform_vec4(self->shader, "color", color);
 
     u32 indices[] = {
-        1, 0, 3, 1, 3, 2, // north (-z)
-        4, 5, 6, 4, 6, 7, // south (+z)
-        5, 1, 2, 5, 2, 6, // east (+x)
-        0, 4, 7, 0, 7, 3, // west (-x)
-        2, 3, 7, 2, 7, 6, // top (+y)
-        5, 4, 0, 5, 0, 1  // bottom (-y)
+            1, 0, 3, 1, 3, 2, // north (-z)
+            4, 5, 6, 4, 6, 7, // south (+z)
+            5, 1, 2, 5, 2, 6, // east (+x)
+            0, 4, 7, 0, 7, 3, // west (-x)
+            2, 3, 7, 2, 7, 6, // top (+y)
+            5, 4, 0, 5, 0, 1  // bottom (-y)
     };
 
     vec3s min = aabb[0], max = aabb[1];
     f32 vertices[] = {
-        min.x, min.y, min.z,
-        max.x, min.y, min.z,
-        max.x, max.y, min.z,
-        min.x, max.y, min.z,
+            min.x, min.y, min.z,
+            max.x, min.y, min.z,
+            max.x, max.y, min.z,
+            min.x, max.y, min.z,
 
-        min.x, min.y, max.z,
-        max.x, min.y, max.z,
-        max.x, max.y, max.z,
-        min.x, max.y, max.z,
+            min.x, min.y, max.z,
+            max.x, min.y, max.z,
+            max.x, max.y, max.z,
+            min.x, max.y, max.z,
     };
 
     vbo_buffer(self->vbo, vertices, 0, (8 * 3) * sizeof(f32));
